fix: integration tests in ci?
also retrying logic for e2e
