+++ test/helpers/apollo-client.ts
@@ -15,9 +15,6 @@ import ws from "ws"
 import { onError } from "@apollo/client/link/error"
 import { baseLogger } from "@services/logger"
 
-import USER_LOGIN from "../e2e/servers/graphql-main-server/mutations/user-login.gql"
-import USER_UPDATE_MUTATION from "../e2e/servers/graphql-main-server/mutations/user-update-username.gql"
-
 export const localIpAddress = "127.0.0.1" as IpAddress
 
 export type ApolloTestClientConfig = {
@@ -121,54 +118,5 @@ export const createApolloClient = (
 export const promisifiedSubscription = (subscription) => {
   return new Promise((resolve, reject) => {
     subscription.subscribe({ next: resolve, error: reject })
-  })
-}
-
-export const loginFromPhoneAndCode = async ({
-  phone,
-  code,
-}: {
-  phone: PhoneNumber
-  code: PhoneCode
-}): Promise<ApolloClient<NormalizedCacheObject>> => {
-  let apolloClient: ApolloClient<NormalizedCacheObject>,
-    disposeClient: () => void = () => null
-
-  let authToken: SessionToken
-  {
-    ;({ apolloClient, disposeClient } = createApolloClient(defaultTestClientConfig()))
-    const input = { phone, code }
-    const result = await apolloClient.mutate({
-      mutation: USER_LOGIN,
-      variables: { input },
-    })
-
-    authToken = result.data.userLogin.authToken
-    expect(authToken).toBeDefined()
-    disposeClient()
-  }
-
-  {
-    // Create a new authenticated client
-    ;({ apolloClient, disposeClient } = createApolloClient(
-      defaultTestClientConfig(authToken),
-    ))
-
-    return apolloClient
-  }
-}
-
-export const updateUsername = async ({
-  username,
-  apolloClient,
-}: {
-  apolloClient: ApolloClient<NormalizedCacheObject>
-  username: Username
-}) => {
-  const input = { username }
-
-  await apolloClient.mutate({
-    mutation: USER_UPDATE_MUTATION,
-    variables: { input },
   })
 }
